Fix: fix the bug that set hotend_offset error

+++ snapmaker/src/module/toolhead_dualextruder.cpp
@@ -920,7 +920,7 @@ ErrCode ToolHeadDualExtruder::HmiSetHotendOffset(SSTP_Event_t &event) {
     goto EXIT;
   }
 
-  tmp_offset = (event.data[1]<<24 | event.data[2]<<16 | event.data[3]<<8 | event.data[4])/1000;
+  tmp_offset = (float)(event.data[1]<<24 | event.data[2]<<16 | event.data[3]<<8 | event.data[4])/1000;
   if (tmp_offset > nozzle_offset[event.data[0]][1] + HOTEND_OFFSET_MAX_DEVIATION || \
       tmp_offset < nozzle_offset[event.data[0]][1] - HOTEND_OFFSET_MAX_DEVIATION) {
     err = E_PARAM;
